Update props naming consistency
Using a naming style like `value-as-integer` for props in the docs comes off a bit misleading when the naming convention in the code is actually in camel case as  `valueAsInteger`. 

This pull request updates occurrences like this in the docs to enforce consistency. I noticed this when I tried copying and pasting the properties directly until I had to dig into the code to really see what naming convention was used.